requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
boto3==1.34.121
botocore==1.34.121
certifi==2024.6.2
charset-normalizer==3.3.2
daily-python==0.9.1
distro==1.9.0
distutils-extra-python==2.39.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.3
langchain-core==0.2.5
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.75
multidict==6.0.5
numpy==1.26.4
openai==1.32.0
orjson==3.10.3
packaging==23.2
portalocker==2.8.2
PyAudio==0.2.14
pydantic==2.7.3
pydantic_core==2.18.4
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
yarl==1.9.4
